+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * display `#t` and `#f` for boolean values in terminal
 * new `interaction-environment` function + `load` use that env by default
 * improve executable
+* user can define string representation (with object that have `toString` lambda)
 ### Bug fixes
 * fix list function [#27](https://github.com/jcubic/lips/issues/27)
 * fix `(begin (gensym))` macro that evaluate to symbol